+++ cg/models/orders/validators/sample_base_validators.py
@@ -1,0 +1,2 @@
+def snake_case(value: str):
+    return value.lower().replace(" ", "_")
